refactor: streamline project fetching and navigation components
